backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
pandas==2.3.0
numpy==1.24.3
requests==2.32.4 
openpyxl==3.1.5
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
dj-database-url==2.1.0 